Added Sparkfun Library, cut down on bloat. Finalized
